test: move custom_sg test to main pr
